[Web CLI] Set dependencies between module descriptors as all-to-all
We are going to reduce the number of cases when the reverse topological
sorting of KLIBs is performed inside the compiler. Because, practically
it's only necessary in a limited number of cases (see KT-65837).

This means that there are no more strict rules on the order of
the resolved `KotlinLibrary`s that are passed on the input of Web CLI.
The only invariant is that stdlib must be the first library in the list.
The order of the remainder is unimportant.

So we should use a more safe "all-depends-on-all" approach when
building `ModuleDescriptor`s instead of the existing
"current-depends-on-all-preceding" approach.

^KT-65837