[FIR] FirPropertyAccessor: replace `hasBody` with `isCustom`
`hasBody` has incorrect semantic in terms of `hasBackingField`
check. Also, it is not correct for accessors from libraries
as they cannot have bodies, but still are divided on custom and
defaults.
Now `KaPropertySymbol#hasBackingField` works more reliably for libraries,
but still has some corner cases:
* a default accessor with an annotation is serialized and deserialized
as a custom one
* custom accessors with the `field` access inside bodies return *false*
as this information is not preserved during serialization

^KT-74777 Fixed