[KMP Parser] Implement full-fledged KMP Kotlin Parser!
The parser is almost straight port of the existing Kotlin parser converted by J2K
and adapted for the new Syntax lib.

Because of that, there are some remaining issues including naming, optimization and other.
They will be fixed later.

Improve testing infrastructure to support for PSI and LightTree modes
It's used for output and performance comparison.

^KT-77352 Fixed
^KT-77144 Fixed