[Test] Add JSpecify CLI test for package level annotations
There are multiple ways that Java source files can be specified to the
compiler, and the different ways have different behaviors when it comes
to package-info.java files. Java files specified as a directory work
correctly, but Java files specified individually do not. Add a test case
which exposes these problems to validate a correct fix.

^KT-77729