[AA] KaExpressionTypeProvider#returnType: don't throw exception for not callables
Currently, we allow calling `returnType` on any `KtDeclaration`.
However, in K2 we throw an exception later if it turns out
that the declaration is not actually callable.

Instead, let's return `Unit` for such declarations
(this is what K1 already does), document it in the KDoc,
and maybe later evolve the API itself or provide an alternative
with a more specific receiver type or contract.

^KT-59857 Fixed